+++ src/config/generalConfig.ts
@@ -1,5 +1,8 @@
+import { verifyCriticalEnvVariable } from "../utils/helpers";
+
+verifyCriticalEnvVariable('JWT_SECRET');
 
 export const generalConfig = {
   jwtSecret: process.env.JWT_SECRET,
-  jwtTtl: process.env.JWT_TTL
+  jwtTtl: process.env.JWT_TTL || '1hr',
 }