+++ src/context.ts
@@ -1,6 +1,6 @@
 import { Logger } from 'pino';
 import { DMPHubAPI } from './datasources/dmphub-api';
-import { MysqlDataSource } from './datasources/mysqlDB';
+import { MySQLDataSource } from './datasources/mySQLDataSource';
 
 export type MyContext = {
   token?: string | undefined;
@@ -8,6 +8,6 @@ export type MyContext = {
 
   dataSources: {
     dmphubAPIDataSource: DMPHubAPI;
-    sqlDataSource: MysqlDataSource;
+    sqlDataSource: MySQLDataSource;
   };
 }