+++ src/services/tokenService.ts
@@ -3,6 +3,7 @@ import { Logger } from 'pino';
 import { formatLogMessage } from '../logger';
 import { User } from '../models/User';
 import { generalConfig } from '../config/generalConfig';
+import { UserRole } from '../types';
 
 export interface JWTToken extends JwtPayload {
   id: number,
@@ -11,12 +12,15 @@ export interface JWTToken extends JwtPayload {
 
 // Generate a JWT Token for the given User
 export const generateToken = (user: User): string => {
-  const payload: JWTToken = {
-    id: user.id,
-    email: user.email,
-    role: user.role.toString()
+  if (generalConfig.jwtSecret && user && user.id && user.email) {
+    const payload: JWTToken = {
+      id: user.id,
+      email: user.email,
+      role: user.role.toString() || UserRole.Researcher,
+    }
+    return jwt.sign(payload, generalConfig.jwtSecret as string, { expiresIn: generalConfig.jwtTtl });
   }
-  return jwt.sign(payload, generalConfig.jwtSecret as string, { expiresIn: generalConfig.jwtTtl });
+  return null;
 };
 
 // Verify the incoming JWT
@@ -24,6 +28,9 @@ export const verifyToken = (token: string, logger: Logger): JWTToken => {
   try {
     return jwt.verify(token, generalConfig.jwtSecret as string) as JWTToken;
   } catch(err) {
-    formatLogMessage(logger, { err });
+    if (logger) {
+      formatLogMessage(logger, { err });
+    }
+    return null;
   }
 };