+++ src/utils/helpers.ts
@@ -1,4 +1,3 @@
-
 // Convert a string into an Array (return the default or an empty array if it is null or undefined)
 export function stringToArray(array: any, delimiter: string = ' ', defaultResponse: string[] = []): string[] {
   if (typeof array === 'string') {
@@ -7,8 +6,28 @@ export function stringToArray(array: any, delimiter: string = ' ', defaultRespon
   return array || defaultResponse;
 }
 
+// Capitalize the first letter of the string.
+export function capitalizeFirstLetter(str: string): string {
+  if (str) {
+    const val = str.trim();
+
+    if (val.length > 0) {
+      return val.charAt(0).toUpperCase() + val.slice(1);
+    }
+    return val;
+  }
+  return '';
+}
+
 // Email address validation
 export function validateEmail(email: string): boolean {
   const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
   return emailRegex.test(email);
+}
+
+// Helper that will log and error and terminate the Node process if a critical env variable is missing.
+export function verifyCriticalEnvVariable(variable: string): void {
+  if (process.env[variable] === undefined) {
+    console.log(Error(`FATAL ERROR: No ${variable} defined in the environment!`));
+  }
 }