+++ test/setup.ts
@@ -1,16 +1,16 @@
-import { ApolloServer } from '@apollo/server';
-import { addMocksToSchema } from '@graphql-tools/mock';
-import { makeExecutableSchema } from '@graphql-tools/schema';
 
-import { MyContext } from '../src/context';
-import { typeDefs } from '../src/schema';
-import { resolvers } from '../src/resolver';
-import { mocks } from '../src/mocks';
+// Always mock out our config files
+jest.mock('../src/config/generalConfig', () => ({
+  generalConfig: {
+    jwtSecret: 'testing',
+    jwtTtl: 5,
+  }
+}));
 
-// Test server using mocks
-export const server = new ApolloServer<MyContext>({
-  schema: addMocksToSchema({
-    schema: makeExecutableSchema({ typeDefs, resolvers }),
-    mocks,
-  }),
-});
+jest.mock('../src/config/oauthConfig', () => ({
+  oauthConfig: {
+    authorizationCodeLifetime: 10,
+    accessTokenLifetime: 30,
+    refreshTokenLifetime: 30,
+  }
+}))